Using icon collection instead of single model for uploading
