+++ lib/assets/javascripts/cartodb/organization/icons/organization_icon_picker_view.js
@@ -1,6 +1,6 @@
 var $ = require('jquery-cdb-v3');
 var _ = require('underscore-cdb-v3');
-var IconModel = require('./organization_icon_model');
+var IconCollection = require('./organization_icon_collection');
 //var cdb = require('cartodb.js-v3');
 
 var icons = [
@@ -52,6 +52,12 @@ module.exports = cdb.core.View.extend({
     }
     this.template = cdb.templates.getTemplate('organization/organization_icon_picker');
     this.orgId = this.options.orgId;
+    this._iconCollection = new IconCollection(
+      null, {
+        orgId: this.orgId
+      }
+    );
+
     //this.model = new cdb.core.Model({ color: this.options.color });
     //this._initBinds();
     this.render();
@@ -80,14 +86,8 @@ module.exports = cdb.core.View.extend({
 
   _onFileSelected: function () {
     var file = this.$("#iconfile").prop('files');
-    debugger;
-    var iconUpload = new IconModel(
-      null, {
-        orgId: this.orgId
-      }
-    );
 
-    iconUpload.save({
+    var newIcon = this._iconCollection.create({
       kind: 'organization_asset',
       resource: file
     }, {
